website/requirements/base.txt
Django>=5.2,<6.0
Pillow>=10.0.0
django-taggit>=5.0.0
django-simple-history>=3.5.0
djangorestframework>=3.15.0
